+++ js/modules/bootstrap-select.js
@@ -15,6 +15,7 @@ $(function(){
         $('.selectpicker').selectpicker({
           style: 'btn-default bs-placeholder',
           size: 'auto',
+          dropupAuto: false, // forces the dropdown the expand "downwards"
         });
     }